Break utilities.py into 2
