+++ src/pykemon_battle/utils/__init__.py
@@ -5,6 +5,9 @@ from .utilities import (
     manually_choose_moveset,
     randomly_choose_moveset,
     choose_first_four_moves_for_now,
+)
+
+from .battle_dynamics import (
     apply_move,
     player_turn_logic,
     enemy_turn_logic,
